pagination logic

+++ src/main/java/com/ideas2it/healthcare/service/ClinicService.java
@@ -84,7 +84,7 @@ public interface ClinicService {
      */
     String deleteClinicById(Integer id);
 
-    Double getTotalPages();
+    Long getTotalPages();
 
-    void setTotalPages(Double totalPages);
+    void setTotalPages(Long totalPages);
 }
